Rename AddGuidepostRef → AddGeneralRef
(and related renamings)
