+++ app/src/main/java/de/westnordost/streetcomplete/quests/QuestsModule.kt
@@ -99,10 +99,10 @@ import de.westnordost.streetcomplete.quests.fire_hydrant_ref.AddFireHydrantRef
 import de.westnordost.streetcomplete.quests.foot.AddProhibitedForPedestrians
 import de.westnordost.streetcomplete.quests.fuel_service.AddFuelSelfService
 import de.westnordost.streetcomplete.quests.general_fee.AddGeneralFee
+import de.westnordost.streetcomplete.quests.general_ref.AddGeneralRef
 import de.westnordost.streetcomplete.quests.grit_bin_seasonal.AddGritBinSeasonal
 import de.westnordost.streetcomplete.quests.guidepost.AddGuidepostEle
 import de.westnordost.streetcomplete.quests.guidepost.AddGuidepostName
-import de.westnordost.streetcomplete.quests.guidepost.AddGuidepostRef
 import de.westnordost.streetcomplete.quests.guidepost_sport.AddGuidepostSports
 import de.westnordost.streetcomplete.quests.hairdresser.AddHairdresserCustomers
 import de.westnordost.streetcomplete.quests.handrail.AddHandrail
@@ -618,7 +618,7 @@ fun getQuestTypeList(
     EE_QUEST_OFFSET + 9 to AddTreeGenus(),
     EE_QUEST_OFFSET + 39 to AddBarrierLocked(),
     EE_QUEST_OFFSET + 26 to AddIsPharmacyDispensing(),
-    EE_QUEST_OFFSET + 42 to AddGuidepostRef(),
+    EE_QUEST_OFFSET + 42 to AddGeneralRef(),
     EE_QUEST_OFFSET + 43 to AddGuidepostName(),
     EE_QUEST_OFFSET + 44 to AddGuidepostEle(),
     EE_QUEST_OFFSET + 30 to AddShelterType(),
