+++ app/src/main/java/de/westnordost/streetcomplete/quests/general_ref/AddGeneralRefForm.kt
@@ -1,19 +1,19 @@
-package de.westnordost.streetcomplete.quests.guidepost
+package de.westnordost.streetcomplete.quests.general_ref
 
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AlertDialog
 import androidx.core.widget.doAfterTextChanged
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.databinding.QuestGuidepostRefBinding
+import de.westnordost.streetcomplete.databinding.QuestGeneralRefBinding
 import de.westnordost.streetcomplete.quests.AbstractOsmQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 import de.westnordost.streetcomplete.util.ktx.nonBlankTextOrNull
 
-class AddGuidepostRefForm : AbstractOsmQuestForm<GuidepostRefAnswer>() {
+class AddGeneralRefForm : AbstractOsmQuestForm<GeneralRefAnswer>() {
 
-    override val contentLayoutResId = R.layout.quest_guidepost_ref
-    private val binding by contentViewBinding(QuestGuidepostRefBinding::bind)
+    override val contentLayoutResId = R.layout.quest_general_ref
+    private val binding by contentViewBinding(QuestGeneralRefBinding::bind)
 
     override val otherAnswers = listOf(
         AnswerItem(R.string.quest_ref_answer_noRef) { confirmNoRef() }
@@ -27,14 +27,14 @@ class AddGuidepostRefForm : AbstractOsmQuestForm<GuidepostRefAnswer>() {
     }
 
     override fun onClickOk() {
-        applyAnswer(GuidepostRef(ref!!))
+        applyAnswer(GeneralRef(ref!!))
     }
 
     private fun confirmNoRef() {
         val ctx = context ?: return
         AlertDialog.Builder(ctx)
             .setTitle(R.string.quest_generic_confirmation_title)
-            .setPositiveButton(R.string.quest_generic_confirmation_yes) { _, _ -> applyAnswer(NoVisibleGuidepostRef) }
+            .setPositiveButton(R.string.quest_generic_confirmation_yes) { _, _ -> applyAnswer(NoVisibleGeneralRef) }
             .setNegativeButton(R.string.quest_generic_confirmation_no, null)
             .show()
     }
