+++ app/src/main/java/de/westnordost/streetcomplete/quests/general_ref/GeneralRefAnswer.kt
@@ -1,0 +1,6 @@
+package de.westnordost.streetcomplete.quests.general_ref
+
+sealed interface GeneralRefAnswer
+
+data class GeneralRef(val ref: String) : GeneralRefAnswer
+object NoVisibleGeneralRef : GeneralRefAnswer
