--- app/src/main/java/de/westnordost/streetcomplete/quests/guidepost/GuidepostRefAnswer.kt
@@ -1,6 +1,0 @@
-package de.westnordost.streetcomplete.quests.guidepost
-
-sealed interface GuidepostRefAnswer
-
-data class GuidepostRef(val ref: String) : GuidepostRefAnswer
-object NoVisibleGuidepostRef : GuidepostRefAnswer
