+++ app/src/main/res/drawable/ic_quest_general_ref.xml
