+++ app/src/main/res/layout/quest_general_ref.xml
